practica_04-Etiquetar-documentos/requirements.txt
scikit-learn>=1.0.0
pandas>=1.4.0
nltk>=3.7
python-docx>=0.8.11
PyPDF2>=3.0.0
joblib>=1.1.0
